hopr_metrics

Module metrics

Contains definitions of metric types.

Structs

MultiCounter

Represents a vector of named monotonic unsigned integer counters. Wrapper for IntCounterVec type

MultiGauge

Represents a vector of gauges with floating point values. Wrapper for GaugeVec type

MultiHistogram

Represents a vector of histograms with floating point values. Wrapper for HistogramVec type

SimpleCounter

Represents a simple monotonic unsigned integer counter. Wrapper for IntCounter type

SimpleGauge

Represents a simple gauge with floating point values. Wrapper for Gauge type

SimpleHistogram

Represents a histogram with floating point values. Wrapper for Histogram type

SimpleTimer

Represents a timer handle.

Functions

gather_all_metrics

Gathers all the global Prometheus metrics.