hopr_metrics

Struct MultiCounter

pub struct MultiCounter { /* private fields */ }

Represents a vector of named monotonic unsigned integer counters. Wrapper for IntCounterVec type

Implementations

impl MultiCounter

pub fn name(&self) -> String

Returns the name of the counter vector given at construction.

impl MultiCounter

pub fn new(name: &str, description: &str, labels: &[&str]) -> Result<Self>

Creates a new vector of integer counters with given name, description and counter labels.

pub fn increment_by(&self, label_values: &[&str], by: u64)

Increments counter with given labels by the given number.

pub fn increment(&self, label_values: &[&str])

Increments counter with given labels by 1.

pub fn reset(&self)

Sets the values to zero.

pub fn get(&self, label_values: &[&str]) -> Option<u64>

Retrieves the value of the specified counter

pub fn labels(&self) -> Vec<&str>

Returns the labels of the counters given at construction.